CSS integrated to website
Created a styles.css file and linked it to all pages within website,
and amended the html code to include classes for use in the css code.
Included css for fonts, font-size, font-weight, and color of several
elements. Adjusted all images to autoscale to 400px width.
Ensured cascade worked correctly.